+++ docs/sphinx_doc/en/source/tutorial/208-distribute.md
@@ -61,7 +61,7 @@ b = AgentB(
 In the Independent Process Mode, we need to start the agent server process on the target machine first.
 When starting the agent server process, you need to specify a model config file, which contains the models which can be used in the agent server, the IP address and port of the agent server process
 For example, start two agent server processes on the two different machines with IP `ip_a` and `ip_b`(called `Machine1` and `Machine2` accrodingly).
-You can run the following code on `Machine1`, and make sure you have put your model config file in `model_config_path_a`. The example model config file instances are located under `examples/model_configs_template`.
+You can run the following code on `Machine1`.Before running, make sure that the machine has access to all models that used in your application, specifically, you need to put your model config file in `model_config_path_a` and set environment variables such as your model API key correctly in `Machine1`. The example model config file instances are located under `examples/model_configs_template`.
 
 ```python
 # import some packages
@@ -87,7 +87,7 @@ server.wait_until_terminate()
 > as_server --host ip_a --port 12001 --model-config-path model_config_path_a
 > ```
 
-And put your model config file accordingly in `model_config_path_b` and run the following code on `Machine2`.
+Then put your model config file accordingly in `model_config_path_b`, set environment variables, and run the following code on `Machine2`.
 
 ```python
 # import some packages
