+++ docs/sphinx_doc/zh_CN/source/tutorial/208-distribute.md
@@ -59,7 +59,7 @@ b = AgentB(
 
 在独立进程模式中，需要首先在目标机器上启动智能体服务器进程，启动时需要提供该服务器能够使用的模型的配置信息，以及服务器的 IP 和端口号。
 例如想要将两个智能体服务进程部署在 IP 分别为 `ip_a` 和 `ip_b` 的机器上（假设这两台机器分别为`Machine1` 和 `Machine2`）。
-你可以先在 `Machine1` 上运行如下代码，运行之前请确保已经将模型配置文件放置在 `model_config_path_a` 位置，模型配置文件样例可参考 `examples/model_configs_template`。
+你可以在 `Machine1` 上运行如下代码。在运行之前请确保该机器能够正确访问到应用中所使用的所有模型。具体来讲，需要将用到的所有模型的配置信息放置在 `model_config_path_a` 文件中，并检查API key 等环境变量是否正确设置，模型配置文件样例可参考 `examples/model_configs_template`。
 
 ```python
 # import some packages
@@ -85,7 +85,7 @@ server.wait_until_terminate()
 > as_server --host ip_a --port 12001  --model-config-path model_config_path_a
 > ```
 
-在 `Machine2` 上运行如下代码，这里同样要确保已经将模型配置文件放置在 `model_config_path_b` 位置。
+在 `Machine2` 上运行如下代码，这里同样要确保已经将模型配置文件放置在 `model_config_path_b` 位置并设置环境变量，从而确保运行在该机器上的 Agent 能够正常访问到模型。
 
 ```python
 # import some packages
